pyproject.toml
[tool.poetry]
name = "arnhem"
version = "0.1.3"
description = "Web application for the Arnhem Postal History project"
authors = ["Jason Heppler <jason@jasonheppler.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
django = "^4.1.7"
django-tailwind = "^3.5.0"
django-htmx = "^1.14.0"
django-extensions = "^3.2.1"
django-environ = "^0.10.0"
humanfriendly = "^10.0"
django-crispy-forms = "^2.0"
crispy-tailwind = "^0.5.0"
django-browser-reload = "^1.7.0"
python-dotenv = "^1.0.0"
pillow = "^9.4.0"
django-admin-interface = "^0.24.2"
django-taggit = "^3.1.0"
taggit-selectize = "^2.11.0"
django-multiselectfield = "^0.1.12"
six = "^1.16.0"
geopy = "^2.3.0"
psycopg2 = "^2.9.7"
django-allauth = "^0.54.0"
argon2-cffi = "^21.3.0"
pytest = "^7.4.0"
django-import-export = "^3.2.0"
django-dbml = "^0.6.0"
daphne = "^4.0.0"
django-tables2 = "^2.6.0"
django-filter = "^23.3"
django-model-import = "^0.7.2"
psycopg2-binary = "^2.9.9"
pandas = "^2.1.2"
termcolor = "^2.4.0"
djlint = "^1.34.1"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
djhtml = "^3.0.6"
pre-commit = "^3.2.0"
django-debug-toolbar = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ==== pytest ====
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = [
    "tests.py",
    "test_*.py",
]
